refactor: webhoo helper
