+++ EMS/client-helper-bundle/src/Helper/Webhook/Webhook.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EMS\ClientHelperBundle\Helper\Webhook;
+
+readonly class Webhook
+{
+    /**
+     * @param mixed[] $data
+     */
+    public function __construct(public string $eventName, public array $data)
+    {
+    }
+}
